+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Delete drink notifications with entries
+- Update drink notifications after edits
 
 ## [1.2.1] - 2026-04-13
 ### Fixed